+++ src/RegExpOutline.ts
@@ -3,6 +3,9 @@ import * as vscode from 'vscode';
 /** 拡張子毎ルール */
 interface RulesExt {
 	ext: string;	// ファイルの拡張子
+	showTOF: boolean;	// アウトラインにTOFを表示するか否か
+	showEOF: boolean;	// アウトラインにEOFを表示するか否か
+	bullets: Array<string>;	// 行頭記号のリスト
 	rules: Array<HeaderRule>;	// 見出しルールのリスト
 }
 
@@ -52,6 +55,37 @@ export class RegExpOutline {
 		} catch (exc: any) {
 			console.error('[regexpOutline] Failed to parse rulesExts as JSON.', rulesExtsStr, exc);
 		}
+		// デフォルト値を設定します
+		for (let rulesExtIdx = 0; rulesExtIdx < rulesExts.length; rulesExtIdx++) {
+			let rulesExt: RulesExt = rulesExts[rulesExtIdx];
+			if (!rulesExt.hasOwnProperty('showTOF')) {
+				rulesExt.showTOF = true;
+			}
+			if (!rulesExt.hasOwnProperty('showEOF')) {
+				rulesExt.showEOF = true;
+			}
+			if (rulesExt.hasOwnProperty('bullets')) {
+				rulesExt.rules = [];
+				for (let bulletIdx = 0; bulletIdx < rulesExt.bullets.length; bulletIdx++) {
+					const rule = {
+						"level": bulletIdx + 1,
+						"format": "^" + rulesExt.bullets[bulletIdx] + "(.+)$",
+						"nameIdx": 1,
+						"detail": "",
+					}
+					rulesExt.rules.push(rule);
+				}
+			}
+			for (let ruleIdx = 0; ruleIdx < rulesExt.rules.length; ruleIdx++) {
+				let rule: HeaderRule = rulesExt.rules[ruleIdx];
+				if (!rule.hasOwnProperty('nameIdx')) {
+					rule.nameIdx = 1;
+				}
+				if (!rule.hasOwnProperty('detail')) {
+					rule.detail = '';
+				}
+			}
+		}
 		return rulesExts;
 	}
 
@@ -68,7 +102,9 @@ export class RegExpOutline {
 			if (!this.document.fileName.endsWith(rulesExt.ext)) {
 				continue;
 			}
-			this.createSymbolFile(0, 'TOF', 'top of file');
+			if (rulesExt.showTOF) {
+				this.createSymbolFile(0, 'TOF', 'top of file');
+			}
 			for (let lineIdx = 0; lineIdx < this.document.lineCount; lineIdx++) {
 				const line: vscode.TextLine = this.document.lineAt(new vscode.Position(lineIdx, 0));
 				const rules: Array<HeaderRule> = rulesExt.rules;
@@ -78,7 +114,9 @@ export class RegExpOutline {
 					}
 				}
 			}
-			this.createSymbolFile(this.document.lineCount - 1, 'EOF', 'end of file');
+			if (rulesExt.showEOF) {
+				this.createSymbolFile(this.document.lineCount - 1, 'EOF', 'end of file');
+			}
 			break;
 		}
 		return this.symbols;