+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "regexpOutline" extension will be documented in this 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.0] - 2024-12-15
+
+- [IMP1] Append keys showTOF and showEOF
+  - for [issue 4](https://github.com/longfish801/regexpOutline/issues/4)
+- [IMP2] Establish default values for nameIdx, detail
+  - for [issue 4](https://github.com/longfish801/regexpOutline/issues/4)
+- [IMP3] Append bullets key
+
 ## [1.0.0] - 2023-09-08
 
 - Initial release